use david's isrequired check
