+++ packages/docs-markdown/src/controllers/metadata/metadata-tree-node.ts
@@ -1,6 +1,6 @@
 import { MarkdownString, ThemeIcon, TreeItem, TreeItemCollapsibleState } from 'vscode';
 import { MetadataSource } from './metadata-source';
-import { MetadataKey } from './metadata-type';
+import { MetadataKey } from './metadata-key';
 import { MetadataCategory } from './metadata-category';
 
 export class MetadataTreeNode extends TreeItem {
